Move BeatmapSetDownloader to osu.Game.Beatmaps.Drawables
